requirements.txt
fastapi[full]
pydantic
vosk
pydub
streaming-form-data
httpx
python-multipart
uvicorn
moviepy
yake
diffusers
torch
aiofiles
accelerate
transformers
Pillow